refactor(runtime): split the spec prototype-chain block out of indexing.rs
CI's 2000-line file cap (scripts/check_file_size.sh, test.yml) rejected this
branch: indexing.rs sat 2 lines under the cap on main and the #9220/#9221 work
pushed it to 2070. Move the spec-level indexed [[Get]] / [[HasProperty]] /
[[Set]] block — the ArrayCustomProto classification, the object-prototype
walks, array_spec_{get,has_index,set} and array_oob_prototype_get — into
`indexing_proto_chain.rs`, declared as a CHILD of `indexing` (the same
#[path] pattern indexing_keyed.rs uses) so every parent-private helper stays
reachable through `super::*` and no visibility widens except the one
pub(super) the parent needs back.

Pure move: the comment-stripped line multiset is identical apart from the
module wiring. The addr-class ratchet baseline is redistributed, not raised —
521 sites before and after — and the new file inherits indexing.rs's existing
GcHeader-probe allowlist entry, exactly as indexing_keyed.rs did at the last
split.

indexing.rs 2070 -> 1603; check_file_size, addr_class_inventory,
shape_descriptor_census, check_thread_locals and check_test_registration all
pass.
