+++ scripts/addr_class_allowlist.txt
@@ -34,6 +34,7 @@ crates/perry-runtime/src/array/generic.rs | * | pre-existing GcHeader probe pred
 crates/perry-runtime/src/array/header.rs | * | pre-existing GcHeader probe predating addr_class; address validated by call-site guards (magnitude/registry/is_valid_obj_ptr) -- migrate to addr_class::try_read_gc_header in a follow-up
 crates/perry-runtime/src/array/indexing.rs | * | pre-existing GcHeader probe predating addr_class; address validated by call-site guards (magnitude/registry/is_valid_obj_ptr) -- migrate to addr_class::try_read_gc_header in a follow-up
 crates/perry-runtime/src/array/indexing_keyed.rs | * | same pre-existing GcHeader probe, moved from indexing.rs by the 2,000-line file split (js_array_set_string_key); migrate to addr_class::try_read_gc_header in a follow-up
+crates/perry-runtime/src/array/indexing_proto_chain.rs | * | same pre-existing GcHeader probe, moved from indexing.rs by the 2,000-line file split (the spec [[Get]]/[[HasProperty]]/[[Set]] prototype-chain block); migrate to addr_class::try_read_gc_header in a follow-up
 crates/perry-runtime/src/array/is_array.rs | * | pre-existing GcHeader probe predating addr_class; address validated by call-site guards (magnitude/registry/is_valid_obj_ptr) -- migrate to addr_class::try_read_gc_header in a follow-up
 crates/perry-runtime/src/array/iter_methods.rs | * | pre-existing GcHeader probe predating addr_class; address validated by call-site guards (magnitude/registry/is_valid_obj_ptr) -- migrate to addr_class::try_read_gc_header in a follow-up
 crates/perry-runtime/src/array/iter_object.rs | * | pre-existing GcHeader probe predating addr_class; address validated by call-site guards (magnitude/registry/is_valid_obj_ptr) -- migrate to addr_class::try_read_gc_header in a follow-up
